+++ src/main/java/de/dhbw_ravensburg/webeng2/backend/controller/BookEntityController.java
@@ -1,0 +1,48 @@
+package de.dhbw_ravensburg.webeng2.backend.controller;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+
+import de.dhbw_ravensburg.webeng2.backend.model.Book;
+import de.dhbw_ravensburg.webeng2.backend.repos.BookRepository;
+import io.swagger.v3.oas.annotations.parameters.RequestBody;
+
+@RestController
+@RequestMapping("/api/books")
+public class BookEntityController {
+
+    @Autowired
+    private BookRepository repository;
+
+    @GetMapping("/")
+    public Page<Book> findBooks() {
+        return repository.findAll(Pageable.ofSize(20));
+    }
+
+    @GetMapping("/{id}")
+    public Book findById(@PathVariable String id) {
+        return repository.findById(id).orElseThrow(() -> new BookNotFoundException());
+    }
+
+    @PutMapping("/{id}")
+    @ResponseStatus(HttpStatus.OK)
+    public Book updateBook(
+            @PathVariable("id") final String id, @RequestBody final Book book) {
+        return book;
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    void onIllegalArgumentException(IllegalArgumentException exception) {
+        // ...
+    }
+}
