+++ src/main/java/de/dhbw_ravensburg/webeng2/backend/controller/BookSearchController.java
@@ -1,5 +1,7 @@
 package de.dhbw_ravensburg.webeng2.backend.controller;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,27 +19,15 @@ import de.dhbw_ravensburg.webeng2.backend.repos.BookRepository;
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
 
 @RestController
-@RequestMapping("/api/books")
-public class BookController {
+@RequestMapping("/api/books/search")
+public class BookSearchController {
 
     @Autowired
     private BookRepository repository;
 
-    @GetMapping("/")
-    public Page<Book> findBooks() {
-        return repository.findAll(Pageable.ofSize(20));
-    }
-
-    @GetMapping("/{id}")
-    public Book findById(@PathVariable String id) {
-        return repository.findById(id).orElseThrow(() -> new BookNotFoundException());
-    }
-
-    @PutMapping("/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public Book updateBook(
-            @PathVariable("id") final String id, @RequestBody final Book book) {
-        return book;
+    @GetMapping("findByName/{name}")
+    public List<Book> findByName(@PathVariable String name) {
+        return repository.findByName(name);
     }
 
     @ExceptionHandler(IllegalArgumentException.class)
